# openapi.yaml
openapi: 3.0.3
info:
  title: Swagger WoW Comanion
  version: 0.1.0
security:
  - oAuthSample: []
...
paths:
...
  /user/registration:
    post:
      tags:
        - User
      summary: Register new account
      operationId: userRegistration
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationParameter'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
...
components:
  securitySchemes:
    oAuthSample:
      type: oauth2
      description: >-
        This API uses OAuth 2 with the implicit grant flow. [More
        info](https://api.example.com/docs/auth)
      flows:
        authorizationCode:
          authorizationUrl: 'https://eu.battle.net/oauth/authorize'
          tokenUrl: 'https://eu.battle.net/oauth/token'
          scopes:
            wow.profile: wow profile
  schemas:
...
    UserDTO:
      type: object
      properties:
        userName:
          type: string
        email:
          type: string
          format: email
        battletag:
          type: string
        blizzardId:
          type: integer
        isAdmin:
          type: boolean
      required:
        - userName
        - email
        - battletag
        - blizzardId
        - isAdmin
...
    UserRegistrationParameter:
      type: object
      properties:
        userName:
          type: string
        email:
          type: string
          format: email
          pattern: >-
            ^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$
      required:
        - userName
        - email
...